Rename to GeneralExternalSettingsProviderService
