+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/Services/BrokeredServicesUtilities.cs
@@ -35,9 +35,6 @@ namespace NuGet.PackageManagement.VisualStudio
         public const string SearchServiceName = "Microsoft.VisualStudio.NuGet.PackageSearchService";
         public const string SearchServiceVersion = "1.0.0";
 
-        public const string ExternalSettingsProviderName = "Microsoft.VisualStudio.NuGet.ExternalSettingsProviderService";
-        public const string ExternalSettingsProviderVersion = "1.0.0";
-
         public static async ValueTask<IServiceBroker> GetRemoteServiceBrokerAsync()
         {
             var serviceBrokerContainer = await ServiceLocator.GetGlobalServiceAsync<SVsBrokeredServiceContainer, IBrokeredServiceContainer>();
